server/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-decouple
python-dotenv